+++ src/InvalidSignatureException.php
@@ -1,0 +1,22 @@
+<?php
+/**
+ * Invalid Signature Exception
+ *
+ * @author    Pronamic <info@pronamic.eu>
+ * @copyright 2005-2021 Pronamic
+ * @license   GPL-3.0-or-later
+ * @package   Pronamic\WordPress\Pay\Gateways\OmniKassa2
+ */
+
+namespace Pronamic\WordPress\Pay\Gateways\OmniKassa2;
+
+/**
+ * Invalid Signature Exception
+ *
+ * @author  Remco Tolsma
+ * @version 2.3.3
+ * @since   2.3.3
+ */
+class InvalidSignatureException extends \Exception {
+
+}